Move some values into utils.ts
